requirements.txt
python-telegram-bot==20.3
matplotlib==3.8.1
python-dotenv==1.0.1
supabase==1.0.0

